requirements.txt
# ── Core API ──────────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
jinja2==3.1.4

# ── NLP / ML ──────────────────────────────────────────────────────────────────
transformers==4.44.2
sentencepiece==0.2.0
protobuf==5.27.2
captum==0.7.0
sentence-transformers==3.0.1
scikit-learn==1.5.2
accelerate==0.34.2
numpy>=1.24,<2.0

# ── Document Ingestion ────────────────────────────────────────────────────────
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
langdetect==1.0.9

# ── Infrastructure & Observability ────────────────────────────────────────────
cachetools==5.4.0
psutil==6.0.0
prometheus-client==0.20.0
python-json-logger==2.0.7

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.0
